backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.30.0
certifi==2025.10.5
click==8.3.0
distro==1.9.0
dotenv==0.9.9
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.11.1
openai==2.6.1
psycopg2==2.9.11
pydantic==2.6.2
pydantic_core==2.16.3
PyJWT==2.9.0
python-dotenv==1.0.0
PyYAML==6.0.3
redis==5.3.1
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
